+++ src/segment_reshape/__init__.py
@@ -18,4 +18,4 @@
 #  along with segment-reshape-qgis-plugin. If not, see <https://www.gnu.org/licenses/>.
 
 
-__version__ = "0.0.3"
+__version__ = "0.0.3.post0"