+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+## Unreleased
+
 ## [0.0.3] - 2022-11-17
 
 - Fix: Fix missing toolbar icon by including resource files in setuptools build